fix(core): move to trash confirm modal missing (#7143)
move `MoveToTrash.ConfirmModal` from `RootAppSidebar ` to `CurrentWorkspaceModals `
